+++ webdev/lib/src/serve/webdev_server.dart
@@ -43,6 +43,8 @@ class WebDevServer {
   /// Can be null if client.js injection is disabled.
   final Dwds dwds;
 
+  final ExpressionCompilerService ddcService;
+
   final String target;
 
   WebDevServer._(
@@ -53,6 +55,7 @@ class WebDevServer {
     this.buildResults,
     bool autoRun, {
     this.dwds,
+    this.ddcService,
   }) {
     if (autoRun) {
       dwds?.connectedApps?.listen((connection) {
@@ -67,6 +70,7 @@ class WebDevServer {
 
   Future<void> stop() async {
     await dwds?.stop();
+    await ddcService?.stop();
     await _server.close(force: true);
     _client?.close();
   }
@@ -98,17 +102,35 @@ class WebDevServer {
 
     var cascade = Cascade();
     var client = http.Client();
-    var assetHandler = proxyHandler(
-        'http://localhost:${options.daemonPort}/${options.target}/',
-        client: client);
+    var assetHandler = proxyHandler('http://localhost:${options.daemonPort}/${options.target}/', client: client);
+    var ddcAssetHandler = proxyHandler('http://localhost:${options.daemonPort}/', client: client);
     Dwds dwds;
+    ExpressionCompilerService ddcService;
     if (options.configuration.enableInjectedClient) {
       var assetReader = ProxyServerAssetReader(
         options.daemonPort,
         logWriter,
         root: options.target,
       );
-      var metadataProvider = MetadataProvider(assetReader, logWriter);
+
+      ddcService = options.configuration.enableExpressionEvaluation
+          ? await ExpressionCompilerService.start(
+              options.configuration.hostname,
+              options.port,
+              options.target,
+              ddcAssetHandler,
+              logWriter,
+              options.configuration.verbose,
+            )
+          : null;
+
+      var metadataProvider = MetadataProvider(
+          assetReader, ddcService?.updateDependencies, logWriter);
+
+      var loadStrategy = BuildRunnerRequireStrategyProvider(
+              assetHandler, options.configuration.reload, metadataProvider)
+          .strategy;
+
       dwds = await Dwds.start(
         hostname: options.configuration.hostname,
         assetReader: assetReader,
@@ -117,21 +139,24 @@ class WebDevServer {
         chromeConnection: () async =>
             (await Chrome.connectedInstance).chromeConnection,
         logWriter: logWriter,
-        loadStrategy: BuildRunnerRequireStrategyProvider(
-                assetHandler, options.configuration.reload, metadataProvider)
-            .strategy,
+        loadStrategy: loadStrategy,
         serveDevTools:
             options.configuration.debug || options.configuration.debugExtension,
         verbose: options.configuration.verbose,
         enableDebugExtension: options.configuration.debugExtension,
         enableDebugging: options.configuration.debug,
         spawnDds: !options.configuration.disableDds,
+        expressionCompiler: ddcService,
       );
       pipeline = pipeline.addMiddleware(dwds.middleware);
       cascade = cascade.add(dwds.handler);
+      cascade = cascade.add(assetHandler);
+      if (ddcService != null) {
+        cascade = cascade.add(ddcService.handler);
+      }
+    } else {
+      cascade = cascade.add(assetHandler);
     }
-
-    cascade = cascade.add(assetHandler);
 
     var hostname = options.configuration.hostname;
     var tlsCertChain = options.configuration.tlsCertChain;
@@ -159,6 +184,7 @@ class WebDevServer {
       filteredBuildResults,
       options.configuration.autoRun,
       dwds: dwds,
+      ddcService: ddcService,
     );
   }
 }